+++ src/main/java/pl/put/trie/Trie.java
@@ -3,17 +3,17 @@ package pl.put.trie;
 //http://ken-soft.com/2012/03/20/java-trie-prefix-tree/
 public class Trie {
 	 
-	private ObjectTrie<Long> trie;
+	private ObjectTrie<Integer> trie;
  
 	public Trie() {
-		trie = new ObjectTrie<Long>(null);
+		trie = new ObjectTrie<Integer>(null);
 	}
  
-	public void insert(Long[] node) {
+	public void insert(Integer[] node) {
 		trie.insert(node);
 	}
  
-	public boolean search(Long[] node) {
+	public boolean search(Integer[] node) {
 		return trie.search(node);
 	}
